worker type cleanup
